mcp_bridge/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
google-api-python-client>=2.116.0
google-auth>=2.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
